Remove some obsolete libraries
`AtmegaScanner` wasn't ever useful outside of the Shortcut port (which has been
dropped from the predecessor of this repo, a good while ago).
`FingerprintUSBHost` is not very reliable, and the only user of it (`HostOS`)
stopped using it. `Kaleidoscope-Focus` was superseded by `FocusSerial`, which is
now in core, so `Kaleidoscope-Focus` is permanently broken due to a symbol
conflict.

For these reasons, lets remove them.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>