model01: Use a modified CDC.cpp
At least for OSX, we need to slow down Serial writes. The easiest way to do that
is to add a delay after each write, and to accomplish that, the easiest is to
modify `CDC.cpp`. Without this, on OSX, we end up writing  too fast (and
changing the baud rate does not help, no matter in which direction I change),
and end up losing data.

This is not a great solution. It's not even a solution, but a workaround. But it
works for now.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>