Removed flag -Os for virtual builds
The virtual hardware is among other purposses meant
to be used for debugging firmware issues.

The flag -Os optimizes out specific information that makes
proper debugging difficult in some situations.

Also, the virtual hardware does not come with restrictions
in terms of PROGMEM. That's why space efficiency is not needed there.

Signed-off-by: Florian Fleissner <florian.fleissner@inpartik.de>