Generate virtual/boards.txt from avr/boards.txt instead of linking them
When building the virtual boards, we want to have the base core as
`keyboardio:arduino`, instead of `arduino:arduino`, because we provide the base
ourselves. To this end, instead of symlinking `avr/boards.txt` and
`virtual/boards.txt`, generate the latter from the former.

Once we have other architectures, such as SAMD, we can concat all the boards.txt
files and do the same `base.core` replacements.

This is a pre-requisite for keyboardio/Kaleidoscope#851.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>