avr/boards: Add a compiler flag for the model01 to set the TWI buffer size
With a recent change in Kaleidoscope, the
`kaleidoscope::driver::keyboardio::Model01Side` driver now uses the twi
implementation also used by the Keyboardio Imago. This sets the default buffer
length to 192, and we need to override that here.

We need the override here, because it is used by `twi.c`, so it needs to be
defined when compiling said file in isolation.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>